requirements.txt
aiobotocore==1.2.2
aiohttp==3.7.4.post0
aioitertools==0.7.1
altair==4.1.0
argon2-cffi==20.1.0
astor==0.8.1
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
base58==2.1.0
bleach==3.3.0
blinker==1.4
boto3==1.16.52
botocore==1.19.52
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
enum-compat==0.0.3
faiss-cpu==1.6.1
filelock==3.0.12
fsspec==0.8.7
future==0.18.2
gdown==3.12.2
idna==2.10
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.5
notebook==6.3.0
numpy==1.18.5
packaging==20.9
pandas==1.1.2
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
prometheus-client==0.9.0
prompt-toolkit==3.0.18
protobuf==3.15.6
ptyprocess==0.7.0
pycparser==2.20
pydeck==0.6.1
Pygments==2.8.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
pyzmq==22.0.3
regex==2021.3.17
requests==2.25.1
s3fs==0.5.1
s3transfer==0.3.6
sacremoses==0.0.43
scikit-learn==0.24.1
scipy==1.6.1
Send2Trash==1.5.0
sentence-transformers==0.3.8
sentencepiece==0.1.95
six==1.15.0
streamlit==0.62.0
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tokenizers==0.8.1rc2
toml==0.10.2
toolz==0.11.1
torch==1.6.0
tornado==6.1
tqdm==4.59.0
traitlets==5.0.5
transformers==3.3.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
validators==0.18.2
-e git+https://github.com/simoroma/vector_engine.git@cf8a87b76b5b2c4c7afae55b6e5f5ecf235115e2#egg=vector_engine
watchdog==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
